+++ packages/ui/src/components/sign-in/sign-in.tsx
@@ -20,7 +20,7 @@ export function SignInComponent() {
   const enabledConnections = getEnabledSocialConnectionsFromEnvironment(clerk?.__unstable__environment);
 
   return (
-    <SignIn.Root exampleMode>
+    <SignIn.Root>
       <Common.Loading>
         {isGlobalLoading => {
           return (
@@ -51,6 +51,7 @@ export function SignInComponent() {
                                     busy={isConnectionLoading}
                                     disabled={isGlobalLoading || isConnectionLoading}
                                     icon={IconComponent ? <IconComponent className='text-base' /> : null}
+                                    textVisuallyHidden={enabledConnections.length > 2}
                                   >
                                     {connection?.name || c.provider}
                                   </Connection.Button>